Static code analysis (part 1)

+++ dynadjust/include/measurement_types/dnagpsbaseline.cpp
@@ -366,7 +366,7 @@ void CDnaGpsBaseline::WriteDNAMsr(std::ofstream* dynaml_stream, const dna_msr_fi
 
 	*dynaml_stream << endl;
 
-	UINT16 pad(dmw.msr_type + dmw.msr_ignore + dmw.msr_inst + dmw.msr_targ1 + dmw.msr_targ2);
+	UINT32 pad(dmw.msr_type + dmw.msr_ignore + dmw.msr_inst + dmw.msr_targ1 + dmw.msr_targ2);
 
 	// X
 	*dynaml_stream << setw(pad) << " ";
